chore: User 클래스 위치 변경으로 인한 수정
